modified previous code for adherence to checkstyle
