+++ src/main/java/seedu/address/logic/commands/ViewCommand.java
@@ -1,14 +1,14 @@
 package seedu.address.logic.commands;
 
-import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
-import seedu.address.model.student.SameStudentIDPredicate;
-import seedu.address.model.student.Student;
+import static java.util.Objects.requireNonNull;
 
 import java.util.List;
 import java.util.ListIterator;
 
-import static java.util.Objects.requireNonNull;
+import seedu.address.logic.CommandHistory;
+import seedu.address.model.Model;
+import seedu.address.model.student.SameStudentIdPredicate;
+import seedu.address.model.student.Student;
 
 /**
  * Lists the previous feedback given to the student identified.
@@ -23,9 +23,9 @@ public class ViewCommand extends Command {
             + "Parameters: KEYWORD\n"
             + "Example: " + COMMAND_WORD + " A0123456Z";
 
-    private final SameStudentIDPredicate predicate;
+    private final SameStudentIdPredicate predicate;
 
-    public ViewCommand(SameStudentIDPredicate predicate) {
+    public ViewCommand(SameStudentIdPredicate predicate) {
         this.predicate = predicate;
     }
 
@@ -43,9 +43,10 @@ public class ViewCommand extends Command {
             }
         }
 
-       return new CommandResult(MESSAGE_FAILURE);
+        return new CommandResult(MESSAGE_FAILURE);
     }
 
+    @Override
     public boolean equals(Object other) {
         return other == this
                 || (other instanceof ViewCommand) // instanceof handles nulls
