+++ src/test/java/seedu/address/logic/commands/ViewCommandTest.java
@@ -1,16 +1,13 @@
 package seedu.address.logic.commands;
 
 import org.junit.Test;
-import org.xml.sax.SAXNotRecognizedException;
-import seedu.address.commons.events.ui.ShowHelpRequestEvent;
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.student.SameStudentIDPredicate;
+import seedu.address.model.student.SameStudentIdPredicate;
 import seedu.address.model.student.Student;
 import seedu.address.testutil.PersonBuilder;
-import seedu.address.ui.testutil.EventsCollectorRule;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -24,8 +21,8 @@ public class ViewCommandTest {
 
     @Test
     public void equals() {
-        SameStudentIDPredicate firstPredicate = new SameStudentIDPredicate("A1234567Z");
-        SameStudentIDPredicate secondPredicate = new SameStudentIDPredicate("B1234567Y");
+        SameStudentIdPredicate firstPredicate = new SameStudentIdPredicate("A1234567Z");
+        SameStudentIdPredicate secondPredicate = new SameStudentIdPredicate("B1234567Y");
 
         ViewCommand firstViewCommand = new ViewCommand(firstPredicate);
         ViewCommand secondViewCommand = new ViewCommand(secondPredicate);
@@ -55,13 +52,13 @@ public class ViewCommandTest {
                 .withStudentId("A0166733Y")
                 .withTags("friends").build();
 
-        assertCommandSuccess(new ViewCommand(new SameStudentIDPredicate("A0166733Y")),
+        assertCommandSuccess(new ViewCommand(new SameStudentIdPredicate("A0166733Y")),
                 model,
                 commandHistory,
                 String.format("%s\n %s\n", ViewCommand.MESSAGE_SUCCESS, expectedStudent),
                 expectedModel);
 
-        assertCommandSuccess(new ViewCommand(new SameStudentIDPredicate("B0123456Z")),
+        assertCommandSuccess(new ViewCommand(new SameStudentIdPredicate("B0123456Z")),
                 model,
                 commandHistory,
                 ViewCommand.MESSAGE_FAILURE,
